requirements.txt
aiocron==2.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
cronsim==2.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.65.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
Telethon==1.39.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
